refactor: organize AppProviders and Layout
